Move MeteorViteConfig.ts and ViteLoadRequest.ts to the src root
