+++ npm-packages/meteor-vite/src/error/MeteorViteError.ts
@@ -1,6 +1,6 @@
 import { inspect } from 'util';
 import MeteorPackage from '../meteor/package/components/MeteorPackage';
-import ViteLoadRequest, { RequestContext } from '../vite/ViteLoadRequest';
+import ViteLoadRequest, { RequestContext } from '../ViteLoadRequest';
 import PackageJson from '../../package.json';
 import pc from 'picocolors';
