+++ npm-packages/meteor-vite/src/meteor/IPC/methods/server-development.ts
@@ -3,10 +3,10 @@ import Path from 'path';
 import { createServer, resolveConfig, ViteDevServer } from 'vite';
 import Logger from '../../../utilities/Logger';
 import MeteorEvents, { MeteorIPCMessage } from '../MeteorEvents';
-import { MeteorViteConfig } from '../../../vite/MeteorViteConfig';
+import { MeteorViteConfig } from '../../../MeteorViteConfig';
 import { MeteorStubs } from '../../../vite';
 import { ProjectJson } from '../../../vite/plugin/MeteorStubs';
-import { RefreshNeeded } from '../../../vite/ViteLoadRequest';
+import { RefreshNeeded } from '../../../ViteLoadRequest';
 import CreateIPCInterface, { IPCReply } from '../interface';
 
 let server: ViteDevServer & { config: MeteorViteConfig };
